requirements.txt
gevent
gunicorn
markdown2
pymysql
django==2.2.6
djangorestframework
markdown 
django-filter
monthdelta
Pillow
xlrd
xlsxwriter
django-markdown-deux
xlsxwriter
tablib==0.14.0
diff_match_patch
xlwt
openpyxl
odfpy
pyyaml
MarkupPy